requirements-win.txt
Django==4.2.16
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-spectacular==0.27.2
pytest==8.3.2
pytest-django==4.8.0
